Move pressButton and sendButtonStates to ButtonsApi, singleton socket client
